% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elr_shiny.R
\name{effectLiteGUI}
\alias{effectLiteGUI}
\title{Shiny interface for effectLite}
\usage{
effectLiteGUI(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Option will be passed on to \code{\link[shiny]{runApp}}}
}
\description{
This function calls a shiny interface for effectLite.
}

