% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{package}
\name{EffectLiteR}
\alias{EffectLiteR}
\alias{EffectLiteR-package}
\title{EffectLiteR}
\description{
Use structural equation modeling to estimate average and conditional effects 
of a treatment variable on an outcome variable, taking into account multiple
continuous and categorical covariates.
}
