% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{elrdata_categorical_items}
\alias{elrdata_categorical_items}
\title{Dataset elrdata_categorical_items.}
\format{
A data frame with 10000 rows and 13 variables.
}
\description{
A simulated dataset for testing measurement models with categorical items:
}
\details{
\itemize{
  \item x. Treatment variable with values 0, 1.
  \item z11. indicator for covariate.
  \item z21. indicator for covariate.
  \item z31. indicator for covariate.
  \item z41. indicator for covariate.
  \item z51. indicator for covariate.
  \item y11. indicator for outcome.
  \item y21. indicator for outcome.
  \item y31. indicator for outcome.
  \item y41. indicator for outcome.
  \item y51. indicator for outcome.
  \item y61. indicator for outcome.
  \item y71. indicator for outcome.
}
}
\keyword{datasets}
