% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{elrdata_kieferetal2024}
\alias{elrdata_kieferetal2024}
\title{Dataset elrdata_kieferetal2024.}
\format{
A data frame with 600 rows and 3 variables.
}
\description{
A simulated dataset for logistic regression from Kiefer, Lugauer, and Mayer (2024):
}
\details{
\itemize{
  \item Y. Outcome variable with values 0, 1.
  \item X. Treatment variable with values 0, 1.
  \item Z. continuous covariate.
}
}
\keyword{datasets}
