% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{elrdata_logreg}
\alias{elrdata_logreg}
\title{Dataset elrdata_logreg.}
\format{
A data frame with 10000 rows and 6 variables.
}
\description{
A simulated dataset for testing logistic regression:
}
\details{
\itemize{
  \item y. Outcome variable with values 0, 1.
  \item x. Treatment variable with values 0, 1.
  \item z1. continuous covariate.
  \item z2. continuous covariate.
  \item k1. categorical covariate.
  \item k2. categorical covariate.
}
}
\keyword{datasets}
