% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elr_shiny.R
\name{elrEffectsGUI}
\alias{elrEffectsGUI}
\title{Shiny interface for elrEffects}
\usage{
elrEffectsGUI(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Option will be passed on to \code{\link[shiny]{runApp}}}
}
\description{
This function calls a shiny interface for elrEffects.
}
