% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elr_read_data.R
\name{elrReadData}
\alias{elrReadData}
\title{Read Data File}
\usage{
elrReadData(
  file,
  name = NULL,
  header = "default",
  sep = "default",
  dec = "default",
  use.value.labels = "default",
  na.strings = "NA"
)
}
\arguments{
\item{file}{Name of the file to read.}

\item{name}{Pure file name (without path to file) to read. 
If \code{file} includes a lengthy path name with many special characters, 
specifying this argument in addition to \code{file} may help the function 
to find the file ending.}

\item{header}{See \code{\link[utils]{read.table}}.}

\item{sep}{See \code{\link[utils]{read.table}}.}

\item{dec}{See \code{\link[utils]{read.table}}.}

\item{use.value.labels}{See \code{\link[foreign]{read.spss}}.}

\item{na.strings}{See \code{\link[foreign]{read.spss}}.}
}
\value{
Object of class \code{"data.frame"}.
}
\description{
Tries to determine the format of the data by the file ending and
chooses the appropriate function to read data. Currently supports .csv,
.dat, .txt, .sav, and .xpt and calls \code{\link[utils:read.table]{read.csv}}, \code{\link[utils:read.table]{read.csv2}},
\code{\link[utils]{read.table}}, \code{\link[foreign]{read.spss}},
\code{\link[foreign]{read.xport}} accordingly. The default values for
arguments depend on the function used to read data.
}
