% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/effectliter.R
\name{effectLite}
\alias{effectLite}
\title{Estimate average and conditional effects}
\usage{
effectLite(y, x, k = NULL, z = NULL, propscore = NULL, control = "0",
  measurement = character(), data, fixed.cell = FALSE,
  missing = "listwise", se = "standard", bootstrap = 1000L,
  syntax.only = FALSE, interactions = "all", ...)
}
\arguments{
\item{y}{Dependent variable (character string).}

\item{x}{Treatment variable (character string).}

\item{k}{Vector of categorical covariates (character vector).}

\item{z}{Vector of continuous covariates (character vector).}

\item{propscore}{Vector of covariates (character vector) that will be used to compute (multiple) propensity scores based on a multinomial regression without interactions. Requires package nnet to be installed. Alternatively, the user can specify a formula with the treatment variable as dependent variable for more control of the propensity score model.}

\item{control}{Value of x that should be used as control group.}

\item{measurement}{Measurement model.}

\item{data}{A data frame.}

\item{fixed.cell}{logical. If FALSE (default), the group sizes are treated as
stochastic rather than fixed.}

\item{missing}{Missing data handling to be used by lavaan::sem(). Will be
passed on to sem()}

\item{se}{Standard error to be used by lavaan::sem(). Will be
passed on to sem().}

\item{bootstrap}{Number of bootstrap drwas, if bootstrapping is used. Will be
passed on to lavaan::sem()-}

\item{syntax.only}{logical. If TRUE, only syntax is returned and the model
will not be estimated.}

\item{interactions}{character. Can be one of c("all","2-way","none") and indicates the type of interaction used in the parameterization of the regression.}

\item{...}{Further arguments passed to lavaan::sem().}
}
\value{
Object of class effectlite.
}
\description{
This function is the main funtion of the package and can be used to estimate
average and conditional effects of a treatment variable on an outcome variable,
taking into account any number of continuous and categorical covariates.
It automatically generates lavaan syntax for a multi-group structural equation
model, runs the model in lavaan, and extracts various average and conditional
effects of interest.
}
\examples{
m1 <- effectLite(y="y", x="x", k="z", control="0", data=nonortho)
print(m1)
}

