% Generated by roxygen2 (4.0.1): do not edit by hand
\name{effectLiteGUI}
\alias{effectLiteGUI}
\title{Shiny interface for effectLite()}
\usage{
effectLiteGUI(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Option will be passed on to shiny::runApp()}
}
\description{
This function calls a shiny interface for effectLite().
}

