% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{example01}
\alias{example01}
\title{Dataset example01.}
\format{
A data frame with 2000 rows and 7 variables.
}
\description{
A simulated dataset. The variables are:
}
\details{
\itemize{
  \item x. Treatment variable with values control, treat1, and treat2.
  \item k1. Categorical covariate with values male and female.
  \item kateg2. Categorical covariate with values 1 and 2.
  \item z1-z3. Continuous covariates.
  \item dv. Coninuous dependent variable.
}
}
\keyword{datasets}
