% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/effectliter.R
\docType{data}
\name{example_multilevel}
\alias{example_multilevel}
\title{Dataset example_multilevel.}
\format{A data frame with 800 rows and 6 variables}
\description{
A simulated dataset. The variables are as follows:
}
\details{
\itemize{
  \item y. coninuous dependent variable
  \item x. treatment variable with values 0, 1
  \item z. continuous covariate
  \item xz. product of x and z
  \item cid. cluster ID
  \item weights. sampling weights
}
}
\keyword{datasets}

