% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{example_multilevel}
\alias{example_multilevel}
\title{Dataset example_multilevel.}
\format{
A data frame with 800 rows and 7 variables.
}
\description{
A simulated dataset with a cluster ID and sampling weights to test multilevel options. The variables are:
}
\details{
\itemize{
  \item y. Coninuous dependent variable.
  \item x. Treatment variable with values 0, 1.
  \item z. Continuous covariate.
  \item xz. Product of x and z.
  \item cid. Cluster ID.
  \item weights. Sampling weights.
  \item iptw. Classic inverse probability of treatment weights based on a logistic regression of x on z. Use with care (only for average effects).
}
}
\keyword{datasets}
